fast_api/requirements.txt
uvicorn 
psycopg2-binary
fastapi
python_dotenv==1.0.1 
